Refactor RunWindow Command Sequence keybindings.
RunWindow Command Sequence editing functions have been changed to take a text_list argument instead of text, due to Python's string immutability. Passing a list allows a function to change values in the parameter and have the caller see those changes in the variable that was passed as a parameter. The inability for a string variable to exhibit these properties resulted in the original majority of RunWindow Command Sequence keybinding code to be done with if elif, rather than the typical dictionary-based keyname-function approach.